requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.32.0
langchain>=0.1.0
langchain-chroma>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
pydantic==2.7.0
pydantic-settings==2.2.0
python-dotenv==1.0.0
chromadb>=0.4.0
PyMuPDF==1.24.0
pdfplumber==0.10.4
Pillow==10.2.0
numpy==1.26.4
easyocr==1.7.0
python-docx==0.8.11
requests==2.31.0
